+++ assets/js/theme/ClickAndBuy.js
@@ -26,23 +26,23 @@ export default class ClickAndBuy extends PageManager {
 
 
         // Recently Viewed Module
-        const recentlyViewed = new Personalization({ 
-                selector: "#recentlyViewedProducts",
-                carousel_selector: ".product-rv-carousel",
-                type: "recentlyViewed",
-                product: {
-                    url: this.context.product.url,
-                    title: this.context.product.title,
-                    product_id: this.context.product.id,
-                    brand: this.context.product.brand.name,
-                    price: this.context.product.price.without_tax.value,
-                    availability: "in stock",
-                    image: replaceSize(this.context.product.main_image.data, 200),
-                    sku: this.context.product.sku
-                }
-            });
+        // const recentlyViewed = new Personalization({ 
+        //         selector: "#recentlyViewedProducts",
+        //         carousel_selector: ".product-rv-carousel",
+        //         type: "recentlyViewed",
+        //         product: {
+        //             url: this.context.product.url,
+        //             title: this.context.product.title,
+        //             product_id: this.context.product.id,
+        //             brand: this.context.product.brand.name,
+        //             price: this.context.product.price.without_tax.value,
+        //             availability: "in stock",
+        //             image: replaceSize(this.context.product.main_image.data, 200),
+        //             sku: this.context.product.sku
+        //         }
+        //     });
 
-        recentlyViewed.init()
+        // recentlyViewed.init()
 
         next()
     }